requirements.txt
requests
beautifulsoup4
click
python-dotenv
pytesseract
Pillow
cloudscraper
lxml
python-telegram-bot==21.10
